If statement to check if captured_key exists

+++ main.py
@@ -14,7 +14,6 @@ blue_led = LED(12)
 vibration_sensor = MotionSensor(25)
 motion_sensor = MotionSensor(26)
 
-
 pwd = str(pathlib.Path(__file__).parent.absolute())
 captured_key_file_location = pwd + '/captured_key.txt'
 
@@ -108,12 +107,13 @@ def scan_code():  # activates the scanner for 5 seconds. Press remote button onc
     else:
         print('Scan failed')
         os.remove(captured_key_file_location)
-        try:
-            shutil.copyfile(backup_file, captured_key_file_location)
-            os.remove(backup_file)
-            print('Restoring backup.')
-        except:
-            pass
+        if os.path.exists(captured_key_file_location):
+            try:
+                shutil.copyfile(backup_file, captured_key_file_location)
+                os.remove(backup_file)
+                print('Restoring backup.')
+            except:
+                pass
 
 
 def set_time_limit(time_object, time_type, time_to_add):  # ads a certain time to an input datetime object (shifts the time by x minutes/seconds).
